Правки в сериализаторе API, вынос фильтров в api.filters.py, правки api.views, правки в api_yamdb.models, правка в reviews.models, правки в setting
