+++ api_yamdb/api_yamdb/models.py
@@ -34,7 +34,11 @@ class YamUser(AbstractUser):
         default=Roles.USER,
 
     )
-    confirmation_code = models.CharField(max_length=100, blank=True, )
+    confirmation_code = models.CharField(
+        max_length=100,
+        blank=True,
+        editable=False,
+        unique=True)
 
     @property
     def is_admin(self):
