+++ api_yamdb/reviews/models.py
@@ -3,6 +3,7 @@ from django.db import models
 
 from api_yamdb.models import YamUser
 
+
 class Category(models.Model):
     name = models.CharField(
         max_length=256,
